Damage recalculations

+++ src/overlays/actors/ovl_Boss_Mo/z_boss_mo.c
@@ -1812,7 +1812,7 @@ void BossMo_CoreCollisionCheck(BossMo* this, PlayState* play) {
                 Actor_PlaySfx(&this->actor, NA_SE_EN_VALVAISA_DAMAGE1);
                 this->actor.colChkInfo.health -= damage;
                 this->hitCount++;
-                if ((s8)this->actor.colChkInfo.health <= 0) {
+                if (this->actor.colChkInfo.health <= 0) {
                     if (((sMorphaTent1->subCamId == SUB_CAM_ID_DONE) && (sMorphaTent2 == NULL)) ||
                         ((sMorphaTent1->subCamId == SUB_CAM_ID_DONE) && (sMorphaTent2 != NULL) &&
                          (sMorphaTent2->subCamId == SUB_CAM_ID_DONE))) {
